Backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.4; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-yasg==1.21.8; python_version >= '3.6'
gunicorn==23.0.0; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
packaging==24.2; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
rest-framework-simplejwt==0.0.2; python_version >= '3.0'
sqlparse==0.5.3; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
django-extensions==4.1; python_version >= '3.9'
django-cors-headers==3.14.0; python_version >= '3.8'
pillow==10.0.0; python_version >= '3.8'
django-filter==23.2; python_version >= '3.8'
drf-spectacular==0.26.1; python_version >= '3.8'
cloudinary==1.36.0; python_version >= '3.8'
pytest==8.3.5; python_version >= '3.8'
pytest-django==4.5.2; python_version >= '3.8'
pytest-factoryboy==2.6.0; python_version >= '3.8'
chapa==0.1.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.8'
whitenoise==6.8.0; python_version >= '3.8'
python-dateutil==2.8.2; python_version >= '3.6'